pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "openai>=1.68.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "sift-stack-py>=0.4.2",
    "streamlit>=1.43.2",
]
